gptchatbot-api/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
Django==5.2.7
django-ninja==1.4.5
docstring_parser==0.17.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.49.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.10.2
numpy==2.3.4
openai==2.7.1
packaging==25.0
pandas==2.3.3
perplexityai==0.20.0
pillow==12.0.0
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pypdf==6.10.0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
streamlit==1.56.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
et-xmlfile==2.0.0
lxml==6.0.3
openpyxl==3.1.5
python-docx==1.2.0
pdf2image==1.17.0
pymupdf==1.27.2.2
pytesseract==0.3.13
PyYAML==6.0.3
Shapely==2.1.2
easyocr==1.7.2
filelock==3.25.2
fsspec==2026.3.0
imageio==2.37.3
lazy-loader==0.5
mpmath==1.3.0
networkx==3.6.1
opencv-python-headless==4.13.0.92
pyclipper==1.4.0
python-bidi==0.6.7
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
sympy==1.14.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
pgvector==0.4.2
pandas
xlrd==2.0.2
